Implement .sequence() trait for Streams
